requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
alsa-lib=1.2.3=h516909a_0
anyio=3.4.0=py38h578d9bd_0
argon2-cffi=21.1.0=py38h497a2fe_2
async_generator=1.10=py_0
attrs=21.2.0=pyhd8ed1ab_0
babel=2.9.1=pyh44b312d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
bleach=4.1.0=pyhd8ed1ab_0
brotli=1.0.9=h7f98852_6
brotli-bin=1.0.9=h7f98852_6
brotlipy=0.7.0=py38h497a2fe_1003
c-ares=1.18.1=h7f98852_0
ca-certificates=2021.10.8=ha878542_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2021.10.8=py38h578d9bd_1
cffi=1.15.0=py38h3931269_0
charset-normalizer=2.0.8=pyhd8ed1ab_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.0=py38h3e25421_0
cudatoolkit=10.1.243=h036e899_9
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.6=h48d8840_2
debugpy=1.5.1=py38h709712a_0
decorator=5.1.0=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
expat=2.4.1=h9c3ff4c_0
fontconfig=2.13.1=hba837de_1005
fonttools=4.28.2=py38h497a2fe_0
freetype=2.10.4=h0708190_1
gettext=0.19.8.1=h73d1719_1008
glib=2.70.1=h780b84a_0
glib-tools=2.70.1=h780b84a_0
gst-plugins-base=1.18.5=hf529b03_2
gstreamer=1.18.5=h9f60fe5_2
h5py=3.6.0=nompi_py38hfbb2109_100
hdf5=1.12.1=nompi_h2750804_102
icu=68.2=h9c3ff4c_0
idna=3.1=pyhd3deb0d_0
importlib-metadata=4.8.2=py38h578d9bd_0
importlib_resources=5.4.0=pyhd8ed1ab_0
ipykernel=6.5.0=py38he5a9106_1
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.5=pyhd8ed1ab_0
jbig=2.1=h7f98852_2003
jedi=0.18.1=py38h578d9bd_0
jinja2=3.0.3=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=4.2.1=pyhd8ed1ab_0
jupyter_client=7.1.0=pyhd8ed1ab_0
jupyter_core=4.9.1=py38h578d9bd_1
jupyter_server=1.12.1=pyhd8ed1ab_0
jupyterlab=3.2.4=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.8.2=pyhd8ed1ab_0
jupyterlab_widgets=1.0.2=pyhd8ed1ab_0
kiwisolver=1.3.2=py38h1fd1430_1
krb5=1.19.2=hcc1bbae_3
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libblas=3.9.0=12_linux64_openblas
libbrotlicommon=1.0.9=h7f98852_6
libbrotlidec=1.0.9=h7f98852_6
libbrotlienc=1.0.9=h7f98852_6
libcblas=3.9.0=12_linux64_openblas
libclang=11.1.0=default_ha53f305_1
libcurl=7.80.0=h2574ce0_0
libdeflate=1.8=h7f98852_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.4.2=h7f98852_5
libgcc-ng=11.2.0=h1d223b6_11
libgfortran-ng=11.2.0=h69a702a_11
libgfortran5=11.2.0=h5c6108e_11
libglib=2.70.1=h174f98d_0
libiconv=1.16=h516909a_0
liblapack=3.9.0=12_linux64_openblas
libllvm11=11.1.0=hf817b99_2
libnghttp2=1.43.0=h812cca2_1
libnsl=2.0.0=h7f98852_0
libogg=1.3.4=h7f98852_1
libopenblas=0.3.18=pthreads_h8fe5266_0
libopus=1.3.1=h7f98852_1
libpng=1.6.37=h21135ba_2
libpq=13.5=hd57d9b9_0
libsodium=1.0.18=h36c2ea0_1
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=he4da1e4_11
libtiff=4.3.0=h6f004c6_2
libuuid=2.32.1=h7f98852_1000
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.2.1=h7f98852_0
libxcb=1.13=h7f98852_1004
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
libzlib=1.2.11=h36c2ea0_1013
llvm-openmp=12.0.1=h4bd325d_1
lz4-c=1.9.3=h9c3ff4c_1
markupsafe=2.0.1=py38h497a2fe_1
matplotlib=3.5.0=py38h578d9bd_0
matplotlib-base=3.5.0=py38hf4fb855_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py38h497a2fe_1005
mkl=2021.4.0=h8d4b97c_729
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=8.0.27=ha770c72_1
mysql-libs=8.0.27=hfa10184_1
nbclassic=0.3.4=pyhd8ed1ab_0
nbclient=0.5.9=pyhd8ed1ab_0
nbconvert=6.3.0=py38h578d9bd_1
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.2=h58526e2_4
nest-asyncio=1.5.1=pyhd8ed1ab_0
ninja=1.10.2=h4bd325d_1
notebook=6.4.6=pyha770c72_0
nspr=4.32=h9c3ff4c_1
nss=3.72=hb5efdd6_0
numpy=1.21.4=py38he2449b9_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1l=h7f98852_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.4=py38h43a58ef_1
pandoc=2.16.2=h7f98852_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.2=pyhd8ed1ab_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.4.0=py38h8e6f84c_0
pip=21.3.1=pyhd8ed1ab_0
prometheus_client=0.12.0=pyhd8ed1ab_0
prompt-toolkit=3.0.22=pyha770c72_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.10.0=pyhd8ed1ab_0
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=3.0.6=pyhd8ed1ab_0
pyqt=5.12.3=py38h578d9bd_8
pyqt-impl=5.12.3=py38h0ffb2e6_8
pyqt5-sip=4.19.18=py38h709712a_8
pyqtchart=5.12=py38h7400c14_8
pyqtwebengine=5.12.1=py38h7400c14_8
pyrsistent=0.18.0=py38h497a2fe_0
pysocks=1.7.1=py38h578d9bd_4
python=3.8.12=hb7a2778_2_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.4.0=py3.8_cuda10.1.243_cudnn7.6.3_0
pytz=2021.3=pyhd8ed1ab_0
pyzmq=22.3.0=py38h2035c66_1
qt=5.12.9=hda022c4_4
readline=8.1=h46c0cb4_0
requests=2.26.0=pyhd8ed1ab_1
scipy=1.7.3=py38h56a6a73_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=59.4.0=py38h578d9bd_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.2.0=py38h578d9bd_2
sqlite=3.37.0=h9cd32fc_0
tbb=2021.4.0=h4bd325d_1
terminado=0.12.1=py38h578d9bd_1
testpath=0.5.0=pyhd8ed1ab_0
tk=8.6.11=h27826a3_1
torchaudio=0.4.0=py38
torchvision=0.5.0=py38_cu101
tornado=6.1=py38h497a2fe_2
tqdm=4.62.3=pyhd8ed1ab_0
traitlets=5.1.1=pyhd8ed1ab_0
unicodedata2=13.0.0.post2=py38h497a2fe_4
urllib3=1.26.7=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=1.2.1=py38h578d9bd_0
wheel=0.37.0=pyhd8ed1ab_1
widgetsnbextension=3.5.2=py38h578d9bd_1
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h516909a_1
zeromq=4.3.4=h9c3ff4c_1
zipp=3.6.0=pyhd8ed1ab_0
zlib=1.2.11=h36c2ea0_1013
zstd=1.5.0=ha95c52a_0
